move spec files to webservice package
