+++ src/test/scala/org/broadinstitute/dsde/firecloud/webservice/ManagedGroupApiServiceSpec.scala
@@ -1,9 +1,9 @@
-package org.broadinstitute.dsde.firecloud.service
+package org.broadinstitute.dsde.firecloud.webservice
 
 import akka.http.scaladsl.model.StatusCodes.{Created, NoContent, OK}
 import akka.http.scaladsl.server.Route.{seal => sealRoute}
 import org.broadinstitute.dsde.firecloud.model.WithAccessToken
-import org.broadinstitute.dsde.firecloud.webservice.ManagedGroupApiService
+import org.broadinstitute.dsde.firecloud.service.{BaseServiceSpec, ManagedGroupService}
 
 import scala.concurrent.ExecutionContext
