+++ src/test/scala/org/broadinstitute/dsde/firecloud/webservice/MethodsApiServiceMultiACLSpec.scala
@@ -1,16 +1,16 @@
-package org.broadinstitute.dsde.firecloud.service
+package org.broadinstitute.dsde.firecloud.webservice
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
-import org.broadinstitute.dsde.firecloud.model.MethodRepository._
-import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol.impMethodAclPair
-import org.broadinstitute.dsde.firecloud.webservice.MethodsApiService
-import org.broadinstitute.dsde.rawls.model.MethodRepoMethod
 import akka.http.scaladsl.model.HttpMethods
 import akka.http.scaladsl.model.StatusCodes._
+import akka.http.scaladsl.server.Route.{seal => sealRoute}
+import org.broadinstitute.dsde.firecloud.model.MethodRepository._
+import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol.impMethodAclPair
+import org.broadinstitute.dsde.firecloud.model.UserInfo
+import org.broadinstitute.dsde.firecloud.service.{AgoraPermissionService, BaseServiceSpec, ServiceSpec}
+import org.broadinstitute.dsde.rawls.model.MethodRepoMethod
 import spray.json.DefaultJsonProtocol._
 import spray.json._
-import akka.http.scaladsl.server.Route.{seal => sealRoute}
-import org.broadinstitute.dsde.firecloud.model.UserInfo
 
 import scala.concurrent.ExecutionContext
